scan require optimizations/fixes
