+++ scan/lib/scan/slack_poster.rb
@@ -1,3 +1,5 @@
+require_relative 'module'
+
 module Scan
   class SlackPoster
     def run(results)
